cdk/cdk-demo/requirements.txt
aws-cdk-lib==2.67.0
constructs>=10.1.260,<11.0.0
boto3==1.26.84
pytest==7.2.2
